[core] Add section to useGridFilter / useGridPage / useGridPageSize / useGridColumnMenu and useGridSelection
