+++ packages/grid/_modules_/grid/hooks/features/columns/useGridColumns.ts
@@ -71,6 +71,9 @@ export function useGridColumns(
     [logger, setGridState, forceUpdate, apiRef],
   );
 
+  /**
+   * API METHODS
+   */
   const getColumn = React.useCallback<GridColumnApi['getColumn']>(
     (field) => gridColumnLookupSelector(apiRef.current.state)[field],
     [apiRef],
@@ -189,7 +192,7 @@ export function useGridColumns(
     [apiRef, logger],
   );
 
-  const colApi: GridColumnApi = {
+  const columnApi: GridColumnApi = {
     getColumn,
     getAllColumns,
     getColumnIndex,
@@ -203,28 +206,11 @@ export function useGridColumns(
     setColumnWidth,
   };
 
-  useGridApiMethod(apiRef, colApi, 'ColApi');
+  useGridApiMethod(apiRef, columnApi, 'GridColumnApi');
 
-  // The effect do not track any value defined synchronously during the 1st render by hooks called after `useGridColumns`
-  // As a consequence, the state generated by the 1st run of this useEffect will always be equal to the initialization one
-  const isFirstRender = React.useRef(true);
-  React.useEffect(() => {
-    if (isFirstRender.current) {
-      isFirstRender.current = false;
-      return;
-    }
-
-    logger.info(`GridColumns have changed, new length ${props.columns.length}`);
-
-    const columnsState = createColumnsState({
-      apiRef,
-      columnsTypes,
-      columnsToUpsert: props.columns,
-      reset: true,
-    });
-    setGridColumnsState(columnsState);
-  }, [logger, apiRef, setGridColumnsState, props.columns, columnsTypes]);
-
+  /**
+   * EVENTS
+   */
   const handlePreProcessorRegister = React.useCallback<
     GridEventListener<GridEvents.preProcessorRegister>
   >(
@@ -259,10 +245,32 @@ export function useGridColumns(
   useGridApiEventHandler(apiRef, GridEvents.preProcessorRegister, handlePreProcessorRegister);
   useGridApiEventHandler(apiRef, GridEvents.viewportInnerSizeChange, handleGridSizeChange);
 
-  // Grid Option Handlers
   useGridApiOptionHandler(
     apiRef,
     GridEvents.columnVisibilityChange,
     props.onColumnVisibilityChange,
   );
+
+  /**
+   * EFFECTS
+   */
+  // The effect do not track any value defined synchronously during the 1st render by hooks called after `useGridColumns`
+  // As a consequence, the state generated by the 1st run of this useEffect will always be equal to the initialization one
+  const isFirstRender = React.useRef(true);
+  React.useEffect(() => {
+    if (isFirstRender.current) {
+      isFirstRender.current = false;
+      return;
+    }
+
+    logger.info(`GridColumns have changed, new length ${props.columns.length}`);
+
+    const columnsState = createColumnsState({
+      apiRef,
+      columnsTypes,
+      columnsToUpsert: props.columns,
+      reset: true,
+    });
+    setGridColumnsState(columnsState);
+  }, [logger, apiRef, setGridColumnsState, props.columns, columnsTypes]);
 }
