+++ packages/grid/_modules_/grid/hooks/features/pagination/useGridPageSize.ts
@@ -40,8 +40,11 @@ export const useGridPageSize = (
     changeEvent: GridEvents.pageSizeChange,
   });
 
-  const setPageSize = React.useCallback(
-    (pageSize: number) => {
+  /**
+   * API METHODS
+   */
+  const setPageSize = React.useCallback<GridPageSizeApi['setPageSize']>(
+    (pageSize) => {
       if (pageSize === gridPageSizeSelector(apiRef.current.state)) {
         return;
       }
@@ -60,18 +63,15 @@ export const useGridPageSize = (
     [apiRef, setGridState, forceUpdate, logger],
   );
 
-  React.useEffect(() => {
-    if (props.pageSize != null && !props.autoPageSize) {
-      apiRef.current.setPageSize(props.pageSize);
-    }
-  }, [apiRef, props.autoPageSize, props.pageSize]);
-
   const pageSizeApi: GridPageSizeApi = {
     setPageSize,
   };
 
   useGridApiMethod(apiRef, pageSizeApi, 'GridPageSizeApi');
 
+  /**
+   * EVENTS
+   */
   const handleUpdateAutoPageSize = React.useCallback(() => {
     const dimensions = apiRef.current.getRootDimensions();
     if (!props.autoPageSize || !dimensions) {
@@ -84,9 +84,18 @@ export const useGridPageSize = (
     apiRef.current.setPageSize(maximumPageSizeWithoutScrollBar);
   }, [apiRef, props.autoPageSize, rowHeight]);
 
+  useGridApiEventHandler(apiRef, GridEvents.viewportInnerSizeChange, handleUpdateAutoPageSize);
+
+  /**
+   * EFFECTS
+   */
+  React.useEffect(() => {
+    if (props.pageSize != null && !props.autoPageSize) {
+      apiRef.current.setPageSize(props.pageSize);
+    }
+  }, [apiRef, props.autoPageSize, props.pageSize]);
+
   React.useEffect(() => {
     handleUpdateAutoPageSize();
   }, [handleUpdateAutoPageSize]);
-
-  useGridApiEventHandler(apiRef, GridEvents.viewportInnerSizeChange, handleUpdateAutoPageSize);
 };
